requirements.txt
amqp==1.4.9
anyjson==0.3.3
apns==2.0.1
asn1crypto==0.23.0
billiard==3.3.0.23
boto==2.44.0
cairocffi==0.8.0
CairoSVG==1.0.22
celery==3.1.23
certifi==2017.7.27.1
cffi==1.7.0
chardet==3.0.4
colorclass==2.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.0.3
cssselect==1.0.1
cssselect2==0.2.1
diff-match-patch==20121119
Django==1.8.9
django-admin-sortable==2.0.1
django-admin-tools==0.8.0
django-allauth==0.23.0
django-appconf==1.0.1
django-celery==3.1.17
django-compressor==2.1
django-cors-headers==1.1.0
django-debug-panel==0.8.3
django-debug-toolbar==1.9.1
django-extra-fields==0.9
django-filter==0.11.0
django-fontawesome==0.3.1
django-gcm==1.0.8
django-grappelli==2.7.2
django-imagekit==3.3
django-import-export==0.2.8
django-jet==1.0.6
django-mailgun==0.8.0
django-modeltranslation==0.12.1
django-nested-admin==3.0.5
django-nested-inline==0.3.5
django-redis-cache==1.7.1
django-rest-swagger==0.3.7
django-reversion==1.9.3
django-sass-processor==0.5.1
django-ses==0.8.0
django-settings-export==1.1.0
django-smart-selects==1.3.1
django-storages==1.4
django-suit==0.2.25
django-summernote==0.8.6
django-super-inlines==0.1.4
django-tastypie==0.12.2
django-videothumbs==0.91
djangorestframework==3.2.3
djrill==1.4.0
docopt==0.6.2
docutils==0.12
enum34==1.1.2
et-xmlfile==1.0.1
fpdf==1.7.2
future==0.16.0
geopy==1.11.0
gevent==1.0.2
googlemaps==2.5.1
greenlet==0.4.7
gunicorn==19.1.1
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
inflect==0.2.5
ipaddress==1.0.16
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
kombu==3.0.35
libsass==0.11.2
lxml==3.8.0
Markdown==2.6.2
MarkupSafe==1.0
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
newrelic==2.56.0.42
num2words==0.5.5
oauthlib==1.0.3
odfpy==1.3.6
olefile==0.44
openapi-codec==1.3.2
openpyxl==2.5.0
packaging==17.1
pdfkit==0.6.1
pdfrw==0.4
pilkit==1.1.13
Pillow==4.3.0
pip-upgrader==1.4.4
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyfcm==1.0.5
PyICU==1.9.8
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
Pyphen==0.9.4
pyserial==3.4
python-dateutil==2.4.2
python-gcm==0.2
python-mimeparse==0.1.4
python-monkey-business==1.0.0
python-openid==2.2.5
python-social-auth==0.2.21
pytz==2015.4
pyxdg==0.25
PyYAML==3.11
razorpay==1.0.0
rcssmin==1.0.6
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.7.0
requests-toolbelt==0.8.0
rjsmin==1.0.12
SecretStorage==2.3.1
simplejson==3.13.2
six==1.10.0
social-auth-core==1.7.0
sqlparse==0.2.4
svg2rlg==0.3
svglib==0.8.1
tablib==0.10.0
terminaltables==3.1.0
tinycss2==0.5
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
uuid==1.30
vine==1.1.4
virtualenv==16.0.0
WeasyPrint==0.39
webencodings==0.5.1
xhtml2pdf==0.0.6
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
